requirements.txt
wikipedia
pyyaml
pyowm